Fix:lgoin bugs ,index mouse event and integrate api auth
