+++ components/Navbar.vue
@@ -13,8 +13,8 @@
       <NuxtLink to="/post-task">
         <v-btn>刊登任務</v-btn>
       </NuxtLink>
-      <v-btn icon size="small"><v-icon>mdi-bell</v-icon>
-        <v-menu activator="parent">
+      <v-btn v-if="_storeAuth.loginToken" icon size="small"><v-icon>mdi-bell</v-icon>
+        <!-- <v-menu activator="parent">
           <v-list>
             <NuxtLink v-for="(item, index) in items" :key="index" :value="index" :to="item.to">
               <v-list-item>
@@ -22,7 +22,7 @@
               </v-list-item>
             </NuxtLink>
           </v-list>
-        </v-menu>
+        </v-menu> -->
       </v-btn>
       <!-- <NuxtLink to="/account"> -->
       <v-btn v-if="_storeAuth.loginToken" icon><v-icon>mdi-account-circle</v-icon>
