+++ services/apis/auth.js
@@ -4,3 +4,17 @@ const { req } = useHttp();
 export const postSignup = (data) => req("POST", "/sign-up", data);
 //登入
 export const postLogin = (data) => req("POST", "/sign-in", data);
+//重設密碼
+export const patchResetPassword = (data, token) =>
+  req("PATCH", "/forgot-reset-password", data, {
+    token: token,
+  });
+//寄送忘記密碼信件
+export const postResetPassword = (data) =>
+  req("POST", "/forgot-password", data);
+//信箱驗證
+export const getVerifyEmail = (token) =>
+  req("GET", `/verify-email`, {}, { token: token });
+//重寄驗證信
+export const postResendVerification = (data) =>
+  req("POST", "/resend-verification", data);
