--- services/apis/resetPassword.js
@@ -1,6 +1,0 @@
-const { req } = useHttp();
-
-export const patchResetPassword = (data, token) =>
-  req("PATCH", "/forgot-reset-password", data, {
-    token: token,
-  });
