--- services/apis/sendResetEmail.js
@@ -1,4 +1,0 @@
-const { req } = useHttp();
-
-export const postResetPassword = (data) =>
-  req("POST", "/forgot-password", data);
