--- services/apis/verifyEmail.js
@@ -1,14 +1,0 @@
-const { req } = useHttp();
-
-export const getVerifyEmail = (token) =>
-  req(
-    "GET",
-    `/verify-email`,
-    {},
-    {
-      token: token,
-    }
-  );
-
-export const postResendVerification = (data) =>
-  req("POST", "/resend-verification", data);
